feat(master): Support setting dp and mp cnt limit for different machines

+++ master/node_selector.go
@@ -68,7 +68,7 @@ type Node interface {
 	PartitionCntLimited() bool
 	IsActiveNode() bool
 	IsWriteAble() bool
-	GetPartitionLimitCnt() uint32
+	GetPartitionLimitCnt() uint64
 	GetTotal() uint64
 	GetUsed() uint64
 	GetAvailableSpace() uint64
